fix: renamed files & adjusted leaflet configuration
